Factorisation de la variable unwanted_array (passage dans les constantes)

+++ app/Http/Controllers/Settings.php
@@ -94,13 +94,7 @@ class Settings extends Controller
     {
         //Déclaration des constantes
         $constants = Config::get('constants');
-
-        $unwanted_array = array(    'Š'=>'S', 'š'=>'s', 'Ž'=>'Z', 'ž'=>'z', 'À'=>'A', 'Á'=>'A', 'Â'=>'A', 'Ã'=>'A', 'Ä'=>'A', 'Å'=>'A', 'Æ'=>'A', 'Ç'=>'C', 'È'=>'E', 'É'=>'E',
-                                'Ê'=>'E', 'Ë'=>'E', 'Ì'=>'I', 'Í'=>'I', 'Î'=>'I', 'Ï'=>'I', 'Ñ'=>'N', 'Ò'=>'O', 'Ó'=>'O', 'Ô'=>'O', 'Õ'=>'O', 'Ö'=>'O', 'Ø'=>'O', 'Ù'=>'U',
-                                'Ú'=>'U', 'Û'=>'U', 'Ü'=>'U', 'Ý'=>'Y', 'Þ'=>'B', 'ß'=>'Ss', 'à'=>'a', 'á'=>'a', 'â'=>'a', 'ã'=>'a', 'ä'=>'a', 'å'=>'a', 'æ'=>'a', 'ç'=>'c',
-                                'è'=>'e', 'é'=>'e', 'ê'=>'e', 'ë'=>'e', 'ì'=>'i', 'í'=>'i', 'î'=>'i', 'ï'=>'i', 'ð'=>'o', 'ñ'=>'n', 'ò'=>'o', 'ó'=>'o', 'ô'=>'o', 'õ'=>'o',
-                                'ö'=>'o', 'ø'=>'o', 'ù'=>'u', 'ú'=>'u', 'û'=>'u', 'ý'=>'y', 'þ'=>'b', 'ÿ'=>'y' );
-
+        $unwanted_array = $constants["UNWANTED_ARRAY"];
         //Récupération des valeurs du formulaire
         $id_projet = $request->input('id_projet'); //l'id projet
         $id_utilisateur = $request->input('id_utilisateur'); //l'id utilisateur
@@ -114,13 +108,21 @@ class Settings extends Controller
             if($action == "delete")
             {
                 //On efface en bdd
-                Droit::where('id_utilisateur', $id_utilisateur)->where('id_projet', $id_projet)->delete();
-                $utilisateur = Utilisateur::where('id', $id_utilisateur)->first();
                 $process = new Process(['../docker/deluser.sh', $projet->port, preg_replace("/[^a-zA-Z0-9]+/", "", strtolower(strtr($utilisateur->prenom, $unwanted_array )).strtolower(strtr($utilisateur->nom, $unwanted_array )))]);
-                $process->run();
+                $return_code = $process->run();
+                if($return_code == 0)
+                {
+                    Droit::where('id_utilisateur', $id_utilisateur)->where('id_projet', $id_projet)->delete();
+                    $utilisateur = Utilisateur::where('id', $id_utilisateur)->first();
+                    //Puis on redirige vers la vue settings
+                    return redirect('settings?id_projet='.$id_projet);
+                }
+                else
+                {
+                    //Redirection vers l'accueil avec l'erreur, en cas d'échec
+                    return redirect('settings?id_projet='.$_GET["id_projet"].'&erreur='.$constants["DOCKER_ERROR"]);
+                }
             }
-            //Puis on redirige vers la vue settings
-            return redirect('settings?id_projet='.$id_projet);
         }
 
         //Récupération des infos de l'utilisateur donné
